+++ Source/JxMutex.cpp
@@ -23,24 +23,31 @@ using namespace Jinx;
 
 #if defined(JINX_WINDOWS)
 static_assert(sizeof(CRITICAL_SECTION) <= MutexDataSize, "Data size must match or exceed CRITICAL_SECTION");
-#else
+#elif defined(JINX_MACOS) || defined(JINX_LINUX)
 static_assert(sizeof(pthread_mutex_t) <= MutexDataSize, "Data size must match or exceed pthread_mutex_t");
+#else
+assert(!"Platform-specific mutex not defined");
 #endif
 
 Mutex::Mutex()
 {
 #if defined(JINX_WINDOWS)
 	InitializeCriticalSection(reinterpret_cast<LPCRITICAL_SECTION>(&m_data));
-#elif defined(OSX) || defined(LINUX)
+#elif defined(JINX_MACOS) || defined(JINX_LINUX)
 	pthread_mutex_init(reinterpret_cast<pthread_mutex_t *>(&m_data), NULL);
+#else
+    assert(!"Platform-specific mutex not defined");
 #endif
 }
+
 Mutex::~Mutex()
 {
 #if defined(JINX_WINDOWS)
 	DeleteCriticalSection(reinterpret_cast<LPCRITICAL_SECTION>(&m_data));
-#elif defined(OSX) || defined(LINUX)
+#elif defined(JINX_MACOS) || defined(JINX_LINUX)
 	pthread_mutex_destroy(reinterpret_cast<pthread_mutex_t *>(&m_data));
+#else
+    assert(!"Platform-specific mutex not defined");
 #endif
 }
 
@@ -48,8 +55,10 @@ void Mutex::lock()
 {
 #if defined(JINX_WINDOWS)
 	EnterCriticalSection(reinterpret_cast<LPCRITICAL_SECTION>(&m_data));
-#elif defined(OSX) || defined(LINUX)
+#elif defined(JINX_MACOS) || defined(JINX_LINUX)
 	pthread_mutex_lock(reinterpret_cast<pthread_mutex_t *>(&m_data));
+#else
+    assert(!"Platform-specific mutex not defined");
 #endif
 }
 
@@ -57,8 +66,10 @@ void Mutex::unlock()
 {
 #if defined(JINX_WINDOWS)
 	LeaveCriticalSection(reinterpret_cast<LPCRITICAL_SECTION>(&m_data));
-#elif defined(OSX) || defined(LINUX)
+#elif defined(JINX_MACOS) || defined(JINX_LINUX)
 	pthread_mutex_unlock(reinterpret_cast<pthread_mutex_t *>(&m_data));
+#else
+    assert(!"Platform-specific mutex not defined");
 #endif
 }